Lighten grest.account_active_stake_cache
Update grest.account_history to remove dependency on grest.account_active_stake_cache
Reduce account_active_stake_cache record maintainance to 4 epochs (only 3 used, unless edge case when synching fresh)
Also retire minimalistic (single-table) functions that are only used once
Minor bug-fixes in grest-poll.sh (echo for dbtip/currtip logging) and dbsync.sh (ensure \${CONFIG}.tmp file is not empty)

+++ scripts/cnode-helper-scripts/dbsync.sh
@@ -68,7 +68,7 @@ check_config_sanity() {
   if [[ "${BYGENHASH}" != "${BYGENHASHCFG}" ]] || [[ "${SHGENHASH}" != "${SHGENHASHCFG}" ]] || [[ "${ALGENHASH}" != "${ALGENHASHCFG}" ]]; then
     cp "${CONFIG}" "${CONFIG}".tmp
     jq --arg BYGENHASH ${BYGENHASH} --arg SHGENHASH ${SHGENHASH} --arg ALGENHASH ${ALGENHASH} '.ByronGenesisHash = $BYGENHASH | .ShelleyGenesisHash = $SHGENHASH | .AlonzoGenesisHash = $ALGENHASH' <"${CONFIG}" >"${CONFIG}".tmp
-    mv -f "${CONFIG}".tmp "${CONFIG}"
+    [[ -s "${CONFIG}".tmp ]] && mv -f "${CONFIG}".tmp "${CONFIG}"
   fi
 }
